Defeated pixel jitter by using vertical sync instead of custom timings for physics and setting movement speeds that align well with 60hz. 120/180/240hz... should also look smooth as butter. Higher than 60hz still need to be tested

+++ src/sprite.h
@@ -20,7 +20,7 @@ class Sprite
 		int offset_y;
 
 	public:
-		int shootingAnimTimer = 0;
+		double shootingAnimTimer = 0;
 		//Initializes the variables
 		Sprite() {
 			current_anim = last_anim = ANIMATION_NONE;
